Make SurrogateFunction an interface and extract the rest of the logic into a base class. It would have been nice to be use it as a mixin but we live with what we have.
